+++ src/main/Recorder.ts
@@ -791,7 +791,12 @@ export default class Recorder extends EventEmitter {
       const obsType = getOBSAudioSourceType(src.type);
 
       console.info('[Recorder] Create audio source', src.id, 'for type', obsType);
+
+      // OBS may have renamed the source if there was a naming conflict,
+      // log that for posterity. Use that name going forward even if it
+      // doesn't match what we asked for.
       const name = noobs.CreateSource(src.id, obsType);
+      console.info('[Recorder] Created audio source', name);
       const settings = noobs.GetSourceSettings(name);
 
       if (src.type === AudioSourceType.PROCESS && src.device) {